% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cookies.r
\name{cookies}
\alias{cookies}
\title{Access cookies in a response.}
\usage{
cookies(x)
}
\arguments{
\item{x}{A response.}
}
\description{
Access cookies in a response.
}
\examples{
r <- GET("http://httpbin.org/cookies/set", query = list(a = 1, b = 2))
cookies(r)
}
\seealso{
\code{\link{set_cookies}()} to send cookies in request.
}

