% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-pool.r
\name{handle_pool}
\alias{handle_pool}
\alias{handle_find}
\alias{handle_reset}
\title{Maintain a pool of handles.}
\format{
An environment.
}
\usage{
handle_pool

handle_find(url)

handle_reset(url)
}
\description{
The handle pool is used to automatically reuse Curl handles for the same
scheme/host/port combination. This ensures that the http session is
automatically reused, and cookies are maintained across requests to a site
without user intervention.
}
\keyword{internal}
