% Generated by roxygen2 (4.0.2): do not edit by hand
\name{has_content}
\alias{has_content}
\title{Does the request have content associated with it?}
\usage{
has_content(x)
}
\description{
Does the request have content associated with it?
}
\examples{
has_content(POST("http://httpbin.org/post", body = FALSE))
has_content(HEAD("http://httpbin.org/headers"))
}
\keyword{internal}

