% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.r
\name{headers}
\alias{headers}
\title{Extract the headers from a response}
\usage{
headers(x)
}
\arguments{
\item{x}{A request object}
}
\description{
Extract the headers from a response
}
\examples{
r <- GET("http://httpbin.org/get")
headers(r)
}
\seealso{
\code{\link{add_headers}()} to send additional headers in a
  request
}

