% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sha1.r
\name{hmac_sha1}
\alias{hmac_sha1}
\title{RSA SHA1}
\usage{
hmac_sha1(key, string)
}
\arguments{
\item{key}{secret key}

\item{string}{data to securely hash}
}
\description{
As described in \url{http://tools.ietf.org/html/rfc3447}.
}
\keyword{internal}

