% Generated by roxygen2 (4.0.0): do not edit by hand
\name{init_oauth1.0}
\alias{init_oauth1.0}
\title{Retrieve OAuth 1.0 access token.}
\usage{
init_oauth1.0(endpoint, app, permission = NULL)
}
\arguments{
\item{endpoint}{An OAuth endpoint, created by \code{\link{oauth_endpoint}}}

\item{app}{An OAuth consumer application, created by
\code{\link{oauth_app}}}

\item{permission}{optional, a string of permissions to ask for.}
}
\description{
See demos for use.
}
\keyword{internal}

