% Generated by roxygen2 (4.0.0): do not edit by hand
\name{new_bin}
\alias{new_bin}
\title{Make bin at \url{http://requestb.in/}}
\usage{
new_bin(browse = TRUE)
}
\arguments{
\item{browse}{if \code{TRUE}, will also open the request bin in your
browser so you can more easily inspect the results.}
}
\description{
Returns the url pointing to a newly created request bin.  This is very
useful for checking that you're sending what you think you're sending.
}
\keyword{internal}

