% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-listener.r
\name{oauth_callback}
\alias{oauth_callback}
\title{The oauth callback url.}
\usage{
oauth_callback()
}
\description{
The url that \code{\link{oauth_listener}} expects that the client be
referred to.
}
\keyword{internal}
