% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{oauth_endpoints}
\alias{oauth_endpoints}
\title{Popular oauth endpoints.}
\format{A named list of objects created by \code{\link{oauth_endpoint}}.}
\usage{
oauth_endpoints
}
\description{
This list provides some common OAuth endpoints.
}
\examples{
oauth_endpoints$twitter
}
\keyword{datasets}

