% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oauth-exchanger.r
\name{oauth_exchanger}
\alias{oauth_exchanger}
\title{Walk the user through the OAuth2 dance without a local webserver.}
\usage{
oauth_exchanger(request_url)
}
\arguments{
\item{request_url}{the url to provide to the user}
}
\description{
This performs a similar function to \code{\link{oauth_listener}},
but without trying do open a browser or start a local webserver.
This manual process can be useful in situations where the user is
remotely accessing the machine outside a browser (say via ssh) or
when it's not possible to successfully receive a callback (such as
when behind a firewall).
}
\details{
This function should generally not be called directly by the user.
}
\keyword{internal}

