% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hmac.r
\name{oauth_sign}
\alias{oauth_sign}
\title{HMAC and RSA SHA1}
\usage{
oauth_sign(key, string, method)
}
\arguments{
\item{key}{secret key}

\item{string}{data to securely sign or hash}
}
\description{
As described in \url{http://datatracker.ietf.org/doc/rfc2104/}.
}
\keyword{internal}

