% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-signature.r
\name{oauth_signature}
\alias{oauth_signature}
\alias{oauth_header}
\title{Generate oauth signature.}
\usage{
oauth_signature(url, method = "GET", app, token = NULL,
  token_secret = NULL, private_key = NULL, other_params = NULL)

oauth_header(info)
}
\arguments{
\item{url, method}{Url and http method of request.}

\item{app}{\code{\link[=oauth_app]{oauth_app()}} object representing application.}

\item{token, token_secret}{OAuth token and secret.}

\item{other_params}{Named argument providing additional parameters
(e.g. \code{oauth_callback} or \code{oauth_body_hash}).}
}
\value{
A list of oauth parameters.
}
\description{
For advanced use only. Occassionally needed for sites that use some
components of the OAuth spec, but not all of them (e.g. 2-legged oauth)
}
\keyword{internal}
