% Generated by roxygen2 (4.0.1): do not edit by hand
\name{progress}
\alias{progress}
\title{Add a progress bar.}
\usage{
progress(type = c("down", "up"))
}
\arguments{
\item{type}{Type of progress to display: either number of bytes uploaded
or downloaded.}
}
\description{
Add a progress bar.
}
\examples{
\donttest{
# If file size is known, you get a progress bar:
x <- GET("http://courses.had.co.nz/12-oscon/slides.zip", progress())
# Otherwise you get the number of bytes downloaded:
x <- GET("http://httpbin.org/drip?numbytes=4000&duration=3", progress())
}
}

