% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-token-utils.R
\name{revoke_all}
\alias{revoke_all}
\title{Revoke all OAuth tokens in the cache.}
\usage{
revoke_all(cache_path = NA)
}
\arguments{
\item{cache_path}{Path to cache file. Defaults to `.httr-oauth` in
current directory.}
}
\description{
Use this function if you think that your token may have been compromised,
e.g. you accidentally uploaded the cache file to github. It's not possible
to automatically revoke all tokens - this function will warn when it can't.
}
