% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.r
\name{set_config}
\alias{set_config}
\alias{reset_config}
\title{Set (and reset) global httr configuration.}
\usage{
set_config(config, override = FALSE)

reset_config()
}
\arguments{
\item{config}{Settings as generated by \code{\link[=add_headers]{add_headers()}},
\code{\link[=set_cookies]{set_cookies()}} or \code{\link[=authenticate]{authenticate()}}.}

\item{override}{if \code{TRUE}, ignore existing settings, if \code{FALSE},
combine new config with old.}
}
\value{
invisibility, the old global config.
}
\description{
Set (and reset) global httr configuration.
}
\examples{
GET("http://google.com")
set_config(verbose())
GET("http://google.com")
reset_config()
GET("http://google.com")
}
\seealso{
Other ways to set configuration:
\code{\link[=config]{config()}},
\code{\link[=with_config]{with_config()}}
}
\concept{ways to set configuration}
