% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_envvar}
\alias{get_envvar}
\alias{set_envvar}
\title{Get and set environment variables.}
\usage{
set_envvar(name, value, scope = c("session", "project", "user", "site"))

get_envvar(name)
}
\arguments{
\item{Name}{of environment variable.}

\item{Value}{of environment variable. Use \code{NA} to unset.}

\item{Scope}{of change. If "session", the change is ephemeral and will
disappear when you restart R. If "project", will modify project the
\code{.Renviron} so it's permanently set for this project/working
directory. If "user", modifies your \code{~/.Renviron} so that it
affects every project. If "site", modifies the site .Renviron so that
it affects every user.}
}
\value{
Invisibly, the previous value of the environment variable.
}
\description{
The problem with this approach is that R doesn't really have any notion of
a "project" .Renviron.
}

