% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sha1.r
\name{sha1_hash}
\alias{sha1_hash}
\title{SHA1 hash}
\usage{
sha1_hash(key, string, method = "HMAC-SHA1")
}
\arguments{
\item{key}{The key to create the hash with}

\item{string}{data to securely hash}

\item{method}{The method to use, either HMAC-SHA1 or RSA-SHA1}
}
\description{
Creates a SHA1 hash of data using either HMAC or RSA.
}
\keyword{internal}

