% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeout.r
\name{timeout}
\alias{timeout}
\title{Set maximum request time.}
\usage{
timeout(seconds)
}
\arguments{
\item{seconds}{number of seconds to wait for a response until giving up.
Can not be less than 1 ms.}
}
\description{
Set maximum request time.
}
\details{
This timeout is passed on to \code{\link[curl:handle_setopt]{curl::handle_setopt()}}.
See there and \code{\link[curl:curl_options]{curl::curl_options()}} for more details.
}
\examples{
\dontrun{
GET("http://httpbin.org/delay/3", timeout(1))
GET("http://httpbin.org/delay/1", timeout(2))
}
}
\seealso{
Other config:
\code{\link[=add_headers]{add_headers()}},
\code{\link[=authenticate]{authenticate()}},
\code{\link[=config]{config()}},
\code{\link[=set_cookies]{set_cookies()}},
\code{\link[=use_proxy]{use_proxy()}},
\code{\link[=user_agent]{user_agent()}},
\code{\link[=verbose]{verbose()}}
}
\concept{config}
