% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.r
\name{use_proxy}
\alias{use_proxy}
\title{Use a proxy to connect to the internet.}
\usage{
use_proxy(url, port = NULL, username = NULL, password = NULL, auth = "basic")
}
\arguments{
\item{url, port}{location of proxy}

\item{username, password}{login details for proxy, if needed}

\item{auth}{type of HTTP authentication to use. Should be one of the
following: basic, digest, digest_ie, gssnegotiate, ntlm, any.}
}
\description{
Use a proxy to connect to the internet.
}
\examples{
# See http://www.hidemyass.com/proxy-list for a list of public proxies
# to test with
# GET("http://had.co.nz", use_proxy("64.251.21.73", 8080), verbose())
}
\seealso{
Other config:
\code{\link[=add_headers]{add_headers()}},
\code{\link[=authenticate]{authenticate()}},
\code{\link[=config]{config()}},
\code{\link[=set_cookies]{set_cookies()}},
\code{\link[=timeout]{timeout()}},
\code{\link[=user_agent]{user_agent()}},
\code{\link[=verbose]{verbose()}}
}
\concept{config}
