% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-agent.r
\name{user_agent}
\alias{user_agent}
\title{Set user agent.}
\usage{
user_agent(agent)
}
\arguments{
\item{agent}{string giving user agent}
}
\description{
Override the default RCurl user agent of \code{NULL}
}
\examples{
\dontrun{
GET("http://httpbin.org/user-agent")
GET("http://httpbin.org/user-agent", user_agent("httr"))
}
}
\seealso{
Other config:
\code{\link[=add_headers]{add_headers()}},
\code{\link[=authenticate]{authenticate()}},
\code{\link[=config]{config()}},
\code{\link[=set_cookies]{set_cookies()}},
\code{\link[=timeout]{timeout()}},
\code{\link[=use_proxy]{use_proxy()}},
\code{\link[=verbose]{verbose()}}
}
\concept{config}
