% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose.r
\name{verbose}
\alias{verbose}
\title{Give verbose output.}
\usage{
verbose(data_out = TRUE, data_in = FALSE, info = FALSE, ssl = FALSE)
}
\arguments{
\item{data_out}{Show data sent to the server.}

\item{data_in}{Show data recieved from the server.}

\item{info}{Show informational text from curl. This is mainly useful
for debugging https and auth problems, so is disabled by default.}

\item{ssl}{Show even data sent/recieved over SSL connections?}
}
\description{
A verbose connection provides much more information about the flow of
information between the client and server.
}
\section{Prefixes}{


\code{verbose()} uses the following prefixes to distinguish between
different components of the http messages:
\itemize{
\item \code{*} informative curl messages
\item \verb{->} headers sent (out)
\item \verb{>>} data sent (out)
\item \verb{*>} ssl data sent (out)
\item \verb{<-} headers received (in)
\item \verb{<<} data received (in)
\item \verb{<*} ssl data received (in)
}
}

\examples{
\dontrun{
GET("http://httpbin.org", verbose())
GET("http://httpbin.org", verbose(info = TRUE))

f <- function() {
  GET("http://httpbin.org")
}
with_verbose(f())
with_verbose(f(), info = TRUE)

# verbose() makes it easy to see exactly what POST requests send
POST_verbose <- function(body, ...) {
  POST("https://httpbin.org/post", body = body, verbose(), ...)
  invisible()
}
POST_verbose(list(x = "a", y = "b"))
POST_verbose(list(x = "a", y = "b"), encode = "form")
POST_verbose(FALSE)
POST_verbose(NULL)
POST_verbose("")
POST_verbose("xyz")
}
}
\seealso{
\code{\link[=with_verbose]{with_verbose()}} makes it easier to use verbose mode
even when the requests are buried inside another function call.

Other config:
\code{\link[=add_headers]{add_headers()}},
\code{\link[=authenticate]{authenticate()}},
\code{\link[=config]{config()}},
\code{\link[=set_cookies]{set_cookies()}},
\code{\link[=timeout]{timeout()}},
\code{\link[=use_proxy]{use_proxy()}},
\code{\link[=user_agent]{user_agent()}}
}
\concept{config}
