% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-function.R
\name{write_function}
\alias{write_function}
\title{S3 object to define response writer.}
\usage{
write_function(subclass, ...)
}
\arguments{
\item{subclass, ...}{Class name and fields. Used in class constructors.}
}
\description{
This S3 object allows you to control how the response body is saved.
}
\keyword{internal}
