% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cache_info}
\alias{cache_info}
\title{Compute caching information for a response.}
\usage{
cache_info(r)
}
\arguments{
\item{r}{A response}
}
\description{
Compute caching information for a response.
}
\examples{
r1 <- GET("https://www.google.com")
cache_info(r1)

r2 <- GET("https://www.google.com/images/srpr/logo11w.png")
cache_info(r2)
}

