% Generated by roxygen2 (4.0.1): do not edit by hand
\name{config}
\alias{config}
\title{Set curl options.}
\usage{
config(...)
}
\arguments{
\item{...}{named Curl options.}
}
\description{
Generally you should only need to use this function to set CURL options
directly if there isn't already a helpful wrapper function, like
\code{\link{set_cookies}}, \code{\link{add_headers}} or
\code{\link{authenticate}}. To use this function effectively requires
some knowledge of CURL, and CURL options. Use \code{\link{httr_options}} to
see a complete list of available options. To see the libcurl documentation
for a given option, use \code{\link{curl_docs}}.
}
\details{
Unlike Curl (and RCurl), all configuration options are per request, not
per handle.
}
\seealso{
\code{\link{set_config}} to set global config defaults, and
 \code{\link{with_config}} to temporarily run code with set options.

Other config: \code{\link{add_headers}};
  \code{\link{authenticate}}; \code{\link{set_cookies}};
  \code{\link{timeout}}; \code{\link{use_proxy}};
  \code{\link{user_agent}}; \code{\link{verbose}}

Other ways to set configuration: \code{\link{reset_config}},
  \code{\link{set_config}}; \code{\link{with_config}},
  \code{\link{with_verbose}}
}

