% Generated by roxygen2 (4.0.2): do not edit by hand
\name{guess_media}
\alias{guess_media}
\title{Guess the media type of a path from its extension.}
\usage{
guess_media(x)
}
\arguments{
\item{x}{path to file}
}
\description{
Guess the media type of a path from its extension.
}
\examples{
guess_media("report.doc")
guess_media("owl.png")
}

