% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.r
\name{has_content}
\alias{has_content}
\title{Does the request have content associated with it?}
\usage{
has_content(x)
}
\description{
Does the request have content associated with it?
}
\examples{
\dontrun{
has_content(POST("http://httpbin.org/post", body = FALSE))
has_content(HEAD("http://httpbin.org/headers"))
}
}
\keyword{internal}
