% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oauth-init.R
\name{init_oauth1.0}
\alias{init_oauth1.0}
\title{Retrieve OAuth 1.0 access token.}
\usage{
init_oauth1.0(endpoint, app, permission = NULL,
  is_interactive = interactive())
}
\arguments{
\item{endpoint}{An OAuth endpoint, created by \code{\link{oauth_endpoint}}}

\item{app}{An OAuth consumer application, created by
\code{\link{oauth_app}}}

\item{permission}{optional, a string of permissions to ask for.}

\item{is_interactive}{Is the current environment interactive?}
}
\description{
See demos for use.
}
\keyword{internal}

