\name{oauth2.0_token}
\alias{oauth2.0_token}
\title{Retrieve access token for}
\usage{
  oauth2.0_token(endpoint, app, scope = NULL)
}
\arguments{
  \item{scope}{a character string of scopes to apply for.}

  \item{endpoint}{An OAuth endpoint, created by
  \code{\link{oauth_endpoint}}}

  \item{app}{An OAuth consumer application, created by
  \code{\link{oauth_app}}}
}
\description{
  Retrieve access token for
}

