% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-endpoint.r
\name{oauth_endpoints}
\alias{oauth_endpoints}
\title{Popular oauth endpoints.}
\usage{
oauth_endpoints(name)
}
\arguments{
\item{name}{One of the following endpoints: linkedin, twitter,
vimeo, google, facebook, github, azure.}
}
\description{
Provides some common OAuth endpoints.
}
\examples{
oauth_endpoints("twitter")
}

