% Generated by roxygen2 (4.0.2): do not edit by hand
\name{oauth_endpoints}
\alias{oauth_endpoints}
\title{Popular oauth endpoints.}
\usage{
oauth_endpoints(name)
}
\arguments{
\item{name}{One of the following endpoints: linkedin, twitter,
vimeo, google, facebook, github.}
}
\description{
Provides some common OAuth endpoints.
}
\examples{
oauth_endpoints("twitter")
}

