% Generated by roxygen2 (4.0.2): do not edit by hand
\name{oauth_service_token}
\alias{oauth_service_token}
\title{Generate OAuth token for service accounts.}
\usage{
oauth_service_token(endpoint, secrets, scope = NULL)
}
\arguments{
\item{endpoint}{An OAuth endpoint, created by \code{\link{oauth_endpoint}}}

\item{secrets}{Secrets loaded from JSON file, downloaded from console.}

\item{scope}{a character vector of scopes to request.}
}
\description{
Generate OAuth token for service accounts.
}
\examples{
\dontrun{
endpoint <- oauth_endpoints("google")
secrets <- jsonlite::fromJSON("~/Desktop/httrtest-45693cbfac92.json")
scope <- "https://www.googleapis.com/auth/bigquery.readonly"

token <- oauth_service_token(endpoint, secret, scope)
}
}

