% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{Add a progress bar.}
\usage{
progress(type = c("down", "up"), con = stdout())
}
\arguments{
\item{type}{Type of progress to display: either number of bytes uploaded
or downloaded.}

\item{con}{Connection to send output too. Usually \code{stdout()} or
\code{stderr}.}
}
\description{
Add a progress bar.
}
\examples{
cap_speed <- config(max_recv_speed_large = 10000)
\donttest{
# If file size is known, you get a progress bar:
x <- GET("http://httpbin.org/bytes/102400", progress(), cap_speed)
# Otherwise you get the number of bytes downloaded:
x <- GET("http://httpbin.org/stream-bytes/102400", progress(), cap_speed)
}
}
