% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe-callback.R
\name{safe_callback}
\alias{safe_callback}
\title{Generate a safe R callback.}
\usage{
safe_callback(f)
}
\arguments{
\item{f}{A function.}
}
\description{
Generate a safe R callback.
}
\keyword{deprecated}

