\name{set_config}
\alias{set_config}
\title{Set global httr configuration.}
\usage{
  set_config(config, override = FALSE)
}
\arguments{
  \item{config}{Settings as generated by
  \code{\link{add_headers}}, \code{\link{set_cookies}} or
  \code{\link{authenticate}}.}

  \item{override}{if \code{TRUE}, ignore existing settings,
  if \code{FALSE}, combine new config with old.}
}
\value{
  invisibility, the old global config.
}
\description{
  Set global httr configuration.
}
\seealso{
  Other ways to set configuration: \code{\link{config}},
  \code{\link{with_config}}
}

