\name{set_config}
\alias{reset_config}
\alias{set_config}
\title{Set (and reset) global httr configuration.}
\usage{
  set_config(config, override = FALSE)

  reset_config()
}
\arguments{
  \item{config}{Settings as generated by
  \code{\link{add_headers}}, \code{\link{set_cookies}} or
  \code{\link{authenticate}}.}

  \item{override}{if \code{TRUE}, ignore existing settings,
  if \code{FALSE}, combine new config with old.}
}
\value{
  invisibility, the old global config.
}
\description{
  Set (and reset) global httr configuration.
}
\examples{
GET("http://google.com")
set_config(verbose())
GET("http://google.com")
reset_config()
GET("http://google.com")
}
\seealso{
  Other ways to set configuration: \code{\link{config}},
  \code{\link{with_config}}
}

