% Generated by roxygen2 (4.0.0): do not edit by hand
\name{set_cookies}
\alias{set_cookies}
\title{Set cookies.}
\usage{
set_cookies(..., .cookies = character(0))
}
\arguments{
\item{...}{a named cookie values}

\item{.cookies}{a named character vector}
}
\description{
Set cookies.
}
\examples{
set_cookies(a = 1, b = 2)
set_cookies(.cookies = c(a = "1", b = "2"))

GET("http://httpbin.org/cookies")
GET("http://httpbin.org/cookies", set_cookies(a = 1, b = 2))
}
\seealso{
Other config: \code{\link{accept_json}},
  \code{\link{accept_xml}}, \code{\link{add_headers}};
  \code{\link{authenticate}}; \code{\link{config}};
  \code{\link{timeout}}; \code{\link{use_proxy}};
  \code{\link{user_agent}}; \code{\link{verbose}}
}

