% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-status.r
\name{stop_for_status}
\alias{message_for_status}
\alias{stop_for_status}
\alias{warn_for_status}
\title{Take action on http error.}
\usage{
stop_for_status(x, message = NULL)

warn_for_status(x, message = NULL)

message_for_status(x, message = NULL)
}
\arguments{
\item{x}{a response, or numeric http code (or other object with
\code{status_code} method)}

\item{message}{The text of the message: either \code{NULL}, a character
  vector or a list.

  If \code{NULL}, the default, will use \code{\link{http_status}} to generate
  the message. If a character vector, all elements will be pasted together.
  If a list, it will be indexed in by the status code.}
}
\value{
If request was successful, the response (invisibly. Otherwise,
  raised a classed http error or warning, as generated by
  \code{\link{http_condition}}
}
\description{
Converts http errors to R errors or warnings - these should always
be used whenever you're creating requests inside a function, so
that the user knows why a request has failed.
}
\examples{
x <- GET("http://httpbin.org/status/200")
stop_for_status(x) # nothing happens
warn_for_status(x)
message_for_status(x)

x <- GET("http://httpbin.org/status/300")
\dontrun{stop_for_status(x)}
warn_for_status(x)
message_for_status(x)

x <- GET("http://httpbin.org/status/404")
\dontrun{stop_for_status(x)}
warn_for_status(x)
message_for_status(x)

# You can provide more information with the message argumgnet
message_for_status(x, "Spreadsheet not found")
message_for_status(x, list(
  "404" = "Spreadsheet not found",
  "403" = "Invalid username/password"
))

}
\seealso{
\code{\link{http_status}} and
  \code{http://en.wikipedia.org/wiki/Http_status_codes} for more information
  on http status codes.

Other response methods: \code{\link{content}},
  \code{\link{http_error}}, \code{\link{http_status}},
  \code{\link{response}}
}

