% Generated by roxygen2 (4.0.1): do not edit by hand
\name{upload_file}
\alias{upload_file}
\title{Upload a file with \code{\link{POST}} or \code{\link{PUT}}.}
\usage{
upload_file(path, type = NULL)
}
\arguments{
\item{path}{path to file}

\item{type}{mime type of path. If not supplied, will be guess by
\code{\link{guess_media}} when needed.}
}
\description{
This is a tiny wrapper for \pkg{RCurl}'s \code{\link[RCurl]{fileUpload}}.
}
\examples{
POST("http://httpbin.org/post",
  body = list(y = upload_file(system.file("CITATION"))))
}

