\name{with_config}
\alias{with_config}
\title{Execute code with configuration set.}
\usage{
  with_config(config = config(), expr, override = FALSE)
}
\arguments{
  \item{expr}{code to execute under specified
  configuration}

  \item{config}{Settings as generated by
  \code{\link{add_headers}}, \code{\link{set_cookies}} or
  \code{\link{authenticate}}.}

  \item{override}{if \code{TRUE}, ignore existing settings,
  if \code{FALSE}, combine new config with old.}
}
\description{
  Execute code with configuration set.
}
\seealso{
  Other ways to set configuration: \code{\link{config}},
  \code{\link{set_config}}
}

