\name{with_config}
\alias{with_config}
\title{Execute code with configuration set.}
\usage{
  with_config(config = config(), expr, override = FALSE)
}
\description{
  Execute code with configuration set.
}
\seealso{
  Other ways to set configuration: \code{\link{config}},
  \code{\link{set_config}}
}

