% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write-function.R
\name{write_function}
\alias{write_function}
\alias{write_init}
\alias{write_opts}
\alias{write_term}
\title{S3 object to define respose writer.}
\usage{
write_function(subclass, ...)

write_init(x)

write_opts(x)

write_term(x)
}
\arguments{
\item{subclass,...}{Class name and fields. Used in class constructors.}

\item{x}{A \code{write_function} object to process.}
}
\description{
This S3 object allows you to control how the response body is saved.
}
\details{
There are three key methods:
\itemize{
  \item \code{write_init()}: called before the write is started. It should
    return a modified object.
  \item \code{write_opts()}: returns a list options passed on to RCurl
  \item \code{write_term()}: called after the request is complete.
    Should return the content (or a pointer to it)
}
}
\keyword{internal}

